Add Mapper

+++ CRUD-example/src/main/java/example/repository/Repository.java
@@ -1,10 +1,12 @@
 package example.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
 
 import example.model.dao.Car;
 
 @org.springframework.stereotype.Repository
-public interface Repository extends JpaRepository<Car, Long> {
+public interface Repository extends JpaRepository<Car,Integer> {
 
 }
